Old quickbms patchers deleted, only new mangetsu based patchers kept. Patcher Readme updated.
