--- tools/patcher/compressonator/README
@@ -1,1 +1,0 @@
-Put compressonator here.
