--- tools/patcher/patch_allui.bat
@@ -1,2 +1,0 @@
-python patch_allui.py
-pause
