--- tools/patcher/quickbms/README
@@ -1,1 +1,0 @@
-Put your quickbms files here.
